remove anti patterns
